ml_new/training/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
torch==2.1.0
transformers==4.36.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.3
openai==1.3.7
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0
toml==0.10.2
aiohttp==3.9.0
python-dotenv==1.1.0